# storage: don't bump tscache on successful CPut/InitPut requests