--- a/pkg/storage/replica_tscache.go
+++ b/pkg/storage/replica_tscache.go
@@ -135,21 +135,19 @@ func (r *Replica) updateTimestampCache(
 			key := keys.TransactionKey(start, pushee.ID)
 			addToTSCache(key, nil, pushee.Timestamp, t.PusherTxn.ID, readCache)
 		case *roachpb.ConditionalPutRequest:
-			if pErr != nil {
-				// ConditionalPut still updates on ConditionFailedErrors.
-				if _, ok := pErr.GetDetail().(*roachpb.ConditionFailedError); !ok {
-					continue
-				}
+			// ConditionalPut only updates on ConditionFailedErrors. On other
+			// errors, no information is returned. On successful writes, the
+			// intent already protects against writes underneath the read.
+			if _, ok := pErr.GetDetail().(*roachpb.ConditionFailedError); ok {
+				addToTSCache(start, end, ts, txnID, true /* readCache */)
 			}
-			addToTSCache(start, end, ts, txnID, true /* readCache */)
 		case *roachpb.InitPutRequest:
-			if pErr != nil {
-				// InitPut still updates on ConditionFailedErrors.
-				if _, ok := pErr.GetDetail().(*roachpb.ConditionFailedError); !ok {
-					continue
-				}
+			// InitPut only updates on ConditionFailedErrors. On other errors,
+			// no information is returned. On successful writes, the intent
+			// already protects against writes underneath the read.
+			if _, ok := pErr.GetDetail().(*roachpb.ConditionFailedError); ok {
+				addToTSCache(start, end, ts, txnID, true /* readCache */)
 			}
-			addToTSCache(start, end, ts, txnID, true /* readCache */)
 		case *roachpb.ScanRequest:
 			resp := br.Responses[i].GetInner().(*roachpb.ScanResponse)
 			if resp.ResumeSpan != nil {