@nvanbenschoten 2019-06-17
> For the allocation numbers in the microbenchmarks, why isn't the allocation size going down more significantly? Was the benchmark simply not running long enough to cycle through a number of 64mb pages or am I missing some reason that makes them matter less than I think?

My guess is that these are allocation counts per operation, so we don't expect to see a huge amount of movement there. We see a reduction of about 100 bytes per write (`change in alloc/op / row count`), which is a reasonable estimate for the size of a tscache node (`len(key) + sizeof(ts) + sizeof(txnid) + sizeof(arenaskl.node)`).